+++ packages/ai/streams/llamaindex-adapter.ts
@@ -1,11 +1,11 @@
+import { convertAsyncIteratorToReadableStream } from '@ai-sdk/provider-utils';
 import { mergeStreams } from '../core/util/merge-streams';
 import { prepareResponseHeaders } from '../core/util/prepare-response-headers';
-import { createStreamDataTransformer, StreamData } from './stream-data';
 import {
-  AIStreamCallbacksAndOptions,
   createCallbacksTransformer,
-  trimStartOfStreamHelper,
-} from './ai-stream';
+  StreamCallbacks,
+} from './stream-callbacks';
+import { createStreamDataTransformer, StreamData } from './stream-data';
 
 type EngineResponse = {
   delta: string;
@@ -13,9 +13,18 @@ type EngineResponse = {
 
 export function toDataStream(
   stream: AsyncIterable<EngineResponse>,
-  callbacks?: AIStreamCallbacksAndOptions,
+  callbacks?: StreamCallbacks,
 ) {
-  return toReadableStream(stream)
+  const trimStart = trimStartOfStream();
+
+  return convertAsyncIteratorToReadableStream(stream[Symbol.asyncIterator]())
+    .pipeThrough(
+      new TransformStream({
+        async transform(message, controller): Promise<void> {
+          controller.enqueue(trimStart(message.delta));
+        },
+      }),
+    )
     .pipeThrough(createCallbacksTransformer(callbacks))
     .pipeThrough(createStreamDataTransformer());
 }
@@ -25,7 +34,7 @@ export function toDataStreamResponse(
   options: {
     init?: ResponseInit;
     data?: StreamData;
-    callbacks?: AIStreamCallbacksAndOptions;
+    callbacks?: StreamCallbacks;
   } = {},
 ) {
   const { init, data, callbacks } = options;
@@ -37,28 +46,21 @@ export function toDataStreamResponse(
   return new Response(responseStream, {
     status: init?.status ?? 200,
     statusText: init?.statusText,
-    headers: prepareResponseHeaders(init, {
+    headers: prepareResponseHeaders(init?.headers, {
       contentType: 'text/plain; charset=utf-8',
       dataStreamVersion: 'v1',
     }),
   });
 }
 
-function toReadableStream(res: AsyncIterable<EngineResponse>) {
-  const it = res[Symbol.asyncIterator]();
-  const trimStartOfStream = trimStartOfStreamHelper();
+function trimStartOfStream(): (text: string) => string {
+  let isStreamStart = true;
 
-  return new ReadableStream<string>({
-    async pull(controller): Promise<void> {
-      const { value, done } = await it.next();
-      if (done) {
-        controller.close();
-        return;
-      }
-      const text = trimStartOfStream(value.delta ?? '');
-      if (text) {
-        controller.enqueue(text);
-      }
-    },
-  });
+  return (text: string): string => {
+    if (isStreamStart) {
+      text = text.trimStart();
+      if (text) isStreamStart = false;
+    }
+    return text;
+  };
 }