+++ packages/cohere/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ai-sdk/cohere
 
+## 1.0.0-canary.4
+
+### Patch Changes
+
+- 3fc69d6: feat (provider/cohere): Pass along tool-plan response content.
+
+## 1.0.0-canary.3
+
+### Patch Changes
+
+- Updated dependencies [8426f55]
+  - @ai-sdk/provider-utils@2.0.0-canary.3
+
+## 1.0.0-canary.2
+
+### Patch Changes
+
+- Updated dependencies [dce4158]
+- Updated dependencies [dce4158]
+  - @ai-sdk/provider-utils@2.0.0-canary.2
+
 ## 1.0.0-canary.1
 
 ### Patch Changes