+++ packages/google/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @ai-sdk/google
 
+## 1.0.0-canary.4
+
+### Major Changes
+
+- 36b03b0: chore (provider/google): remove topK model setting
+- 277fc6b: chore (provider/google): remove facade
+
+## 1.0.0-canary.3
+
+### Patch Changes
+
+- Updated dependencies [8426f55]
+  - @ai-sdk/provider-utils@2.0.0-canary.3
+
+## 1.0.0-canary.2
+
+### Patch Changes
+
+- Updated dependencies [dce4158]
+- Updated dependencies [dce4158]
+  - @ai-sdk/provider-utils@2.0.0-canary.2
+
 ## 1.0.0-canary.1
 
 ### Major Changes