+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @ai-sdk/vue
 
+## 1.0.0-canary.6
+
+### Patch Changes
+
+- Updated dependencies [70f28f6]
+  - @ai-sdk/ui-utils@1.0.0-canary.6
+
+## 1.0.0-canary.5
+
+### Patch Changes
+
+- Updated dependencies [9f81e66]
+- Updated dependencies [8426f55]
+  - @ai-sdk/ui-utils@1.0.0-canary.5
+  - @ai-sdk/provider-utils@2.0.0-canary.3
+
+## 1.0.0-canary.4
+
+### Major Changes
+
+- ca3e586: chore (ui): remove experimental_useAssistant export
+
+### Patch Changes
+
+- Updated dependencies [dce4158]
+- Updated dependencies [dce4158]
+  - @ai-sdk/provider-utils@2.0.0-canary.2
+  - @ai-sdk/ui-utils@1.0.0-canary.4
+
 ## 1.0.0-canary.3
 
 ### Patch Changes